requirements/all.txt
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==23.1.0
autocommand==2.2.2
bandit==1.7.5
cfgv==3.3.1
charset-normalizer==3.1.0
colorama==0.4.6
comtypes==1.1.14
coverage==7.2.3
distlib==0.3.6
filelock==3.11.0
flake8==6.0.0
flake8-bugbear==23.3.23
frozenlist==1.4.1
gitdb==4.0.10
GitPython==3.1.31
identify==2.5.22
idna==3.4
inflect==6.0.4
iniconfig==2.0.0
irc==20.1.0
isort==5.12.0
jaraco.collections==4.1.0
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.logging==3.1.2
jaraco.stream==3.0.3
jaraco.text==3.11.1
json-source-map==1.0.5
jsonschema==4.17.3
keyboard==0.13.5
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
multidict==6.0.5
mypy==1.2.0
mypy-extensions==1.0.0
nodeenv==1.7.0
obs-websocket-py==1.0
packaging==23.1
pbr==5.11.1
pefile==2023.2.7
platformdirs==3.2.0
pluggy==1.0.0
pre-commit==3.2.2
pycodestyle==2.10.0
pycryptodomex==3.17
pydantic==1.10.7
pydirectinput-rgx==2.0.7
pyflakes==3.0.1
pygame==2.6.0
Pygments==2.15.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pysimplegui==5.0.6
psutil==5.9.5
pypiwin32==223
pyqt6==6.7.1
pyrsistent==0.19.3
pytest==7.3.1
pytest-cov==4.0.0
pytest-mock==3.10.0
pytest-random-order==1.1.0
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3
pyuac==0.0.3
pywin32==306
pywin32-ctypes==0.2.1
PyYAML==6.0.1
requests==2.30.0
rich==13.3.4
six==1.16.0
smmap==5.0.0
stevedore==5.0.0
tempora==5.2.2
twitchAPI==3.10.0
types-colorama==0.4.15.11
types-jsonschema==4.17.0.7
types-keyboard==0.13.2.6
typing_extensions==4.5.0
vgamepad==0.0.8
virtualenv==20.21.0
yarl==1.9.4
